% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FITgraphicsFunctions.R
\name{importFit}
\alias{importFit}
\title{Import FIT File}
\usage{
importFit(fileName, param2)
}
\arguments{
\item{fileName}{- whatever the first param is}

\item{parameter2}{- whatever the second param is}
}
\description{
Checks for the existence of a file in the given location and
if it exist, collects data and outputs a Data Frame containing all data in
the file.
}
