% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FITgraphicsFunctions.R
\name{mapFit}
\alias{mapFit}
\title{mapFit}
\usage{
mapFit(fitFile)
}
\arguments{
\item{fitFile}{- {Dataframe} (Required) Must be the output of \code{gitFit()}}
}
\value{
A map showing the GPS route that was taken during the activity.
}
\description{
Plots the map and the path of the activity.
}
