% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FITgraphicsFunctions.R
\name{plotFit}
\alias{plotFit}
\title{plotFit}
\usage{
plotFit(fitFile, varName, showAverage = FALSE)
}
\arguments{
\item{fitFile}{- {Dataframe} (Required) Must be the output of \code{gitFit()}}

\item{varName}{- {Character} (Required) Name of requested variable to be graphed}

\item{showAverage}{{Logical} (Optional)Show average line and value, \code{default = False}}
}
\value{
Returns a plot of the \code{varName} vs time. This The average and
maximum values will be shown based on the input parameters. Each graph will
have a different graph color based on what \code{varName} is chosen.
}
\description{

}
