% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FITgraphicsFunctions.R
\name{getFit}
\alias{getFit}
\title{Import FIT File}
\usage{
getFit(fileName)
}
\arguments{
\item{fileName}{- {Character} (Required) File path corresponding to the FIT file. File must be .fit extension.}
}
\value{
Returns a \code{data.frame} ordered based on the timestamp of each instance.
Each column is named and represents the different variables e.g. heart_rate.
}
\description{
Checks for the existence of a file in the given location and
if it exist, collects data and outputs a \code{data.frame} containing all data in
the file.
}
\examples{

#file_path <- system.file("extdata", "TestRun.fit", package = "FITgraphics")

#loaded_File <- getFit(file_path)

#loaded_File
}
