% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FITgraphicsFunctions.R
\name{plotFit}
\alias{plotFit}
\title{plotFit}
\usage{
plotFit(fitFile, varName)
}
\arguments{
\item{fitFile}{- {Dataframe} Required parameters: data, records}

\item{varName}{- {Character}}
}
\value{

}
\description{

}
