% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FITgraphicsFunctions.R
\name{plotFit}
\alias{plotFit}
\title{plotFit}
\usage{
plotFit(fitFile, varName, showAverage = FALSE)
}
\arguments{
\item{fitFile}{- {Dataframe} Required parameters: data, records}

\item{varName}{- {Character} Name of requested variable to be graphed}

\item{showAverage}{{Logical} Show average line and value, default = False}
}
\value{

}
\description{

}
