% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FITgraphicsFunctions.R
\name{showTrainingEffect}
\alias{showTrainingEffect}
\title{showTrainingEffect}
\usage{
showTrainingEffect(fitFile, AnaerobicTE = FALSE)
}
\arguments{
\item{fitFile}{- Must be the output from the getFit() function in this
package.}

\item{AnaerobicTE}{- (Logical) Default = FALSE. If true, the function will
return the Anaerobic Training Effect rather than the Aerobic Training
Effect.}
}
\value{

}
\description{

}
